fixed TaxService logic , optimised services ,changed entities property types,deleted bank account reference to its owner , added OperationType and TransactionType ,added Pessimistic locks , fixed tests

+++ src/main/java/com/example/pyramid/model/Tax.java
@@ -1,13 +1,14 @@
 package com.example.pyramid.model;
 
 import javax.persistence.Entity;
+import java.math.BigDecimal;
 
 @Entity
 public class Tax extends _BaseEntity{
 
     private String name;
 
-    private Double taxAmount;
+    private BigDecimal taxAmount;
 
     public String getName() {
         return name;
@@ -17,11 +18,11 @@ public class Tax extends _BaseEntity{
         this.name = name;
     }
 
-    public Double getTaxAmount() {
+    public BigDecimal getTaxAmount() {
         return taxAmount;
     }
 
-    public void setTaxAmount(Double taxAmount) {
+    public void setTaxAmount(BigDecimal taxAmount) {
         this.taxAmount = taxAmount;
     }
 }
